refactor(tts): derive OuteTTS helpers from upstream at build time (no hand-copy)
Addresses review feedback on PR #268: the TTS native pipeline reused llama.cpp's
tools/tts/tts.cpp by hand-copying its DSP/prompt/text helpers and default-speaker
strings into tts_dsp.hpp + tts_engine.cpp — a DRY/maintenance hazard that would
silently diverge on every llama.cpp upgrade, and a missing-attribution concern.

tts.cpp cannot simply be added to target_sources: it defines its own main()
(link clash, same reason server.cpp is excluded) and every helper is `static`
(internal linkage — unreachable from another TU). So instead of copying, the
helpers are now DERIVED MECHANICALLY from the pinned upstream source at configure
time:

- cmake/generate-tts-upstream.cmake reads the pinned tools/tts/tts.cpp, keeps the
  pre-main() span, strips `static` from the helpers the engine calls (external
  linkage), and extracts the two default-speaker literals out of main() into
  `extern const` strings. Emits build/tts_generated/tts_upstream_gen.cpp (never
  committed; regenerated from whatever tts.cpp the GIT_TAG resolves to, so a
  version bump is picked up automatically).
- CMakeLists runs it after FetchContent_MakeAvailable(llama.cpp) and compiles the
  generated TU into jllama.
- tts_upstream.h: committed, hand-written declarations of the extracted symbols
  (interface only). tts_engine.cpp keeps only our orchestration + the in-memory
  WAV writer (tts_wav.hpp, ours). tts_dsp.hpp and all copied helpers are removed.

Fail-loud on drift (same contract as patches/): the generator asserts the
`int main(` anchor, every de-static signature, and both speaker literals; a
rename aborts the configure, a type change fails the link. Silent divergence is
impossible.

Bonus: using upstream's real process_text (which calls replace_numbers_with_words)
fixes the previous digit-drop limitation — English numbers are now spoken.

Verified: jllama builds + links, 454 C++ tests pass, and TtsIntegrationTest
synthesizes a valid 24 kHz WAV end-to-end against the real OuteTTS + WavTokenizer
models.

test_tts_dsp.cpp -> test_tts_wav.cpp (now covers only our WAV writer; the DSP is
upstream's, covered end-to-end by TtsIntegrationTest).
